rename EntitiesCollectionApi module filename
